Added new javascript
Added JavaScript that changed the list array to an array of objects, thus dynamically putting an image and p tag into the hotspot annotation classes. Added images in the files to be used in the array. Added styling for the images and hotspotannotation boxes to work on desktop and mobile.